+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Menu item to clear recent files.
+- Menu item to close window. (Linux)
+- Menu item to quit application. (Linux)
 
 ### Fixed